add all test files

+++ src/main/java/com/piseth/java/school/phoneshopenight/service/ModelService.java
@@ -6,5 +6,6 @@ import com.piseth.java.school.phoneshopenight.entity.Model;
 
 public interface ModelService {
 	Model save(Model model);
-	List<Model> getByBrand(Integer brandId);
+	List<Model> getByBrand(Long brandId);
+	
 }
