+++ src/main/java/com/piseth/java/school/phoneshopenight/service/impl/ModelServiceImpl.java
@@ -21,7 +21,7 @@ public class ModelServiceImpl implements ModelService {
 	}
 
 	@Override
-	public List<Model> getByBrand(Integer brandId) {
+	public List<Model> getByBrand(Long brandId) {
 		return modelRepository.findByBrandId(brandId);
 	}
